// src/app/api/models/tenant-admin.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Address } from '../models/address';
import { Tenant } from '../models/tenant';
import { UserDto } from '../models/user-dto';

/**
 * Detalles del TenantAdmin a crear.
 */
export interface TenantAdmin {

  /**
   * Objeto Dirección Completo (Para creación inline)
   */
  address?: Address;

  /**
   * Nivel de Administrador
   */
  adminLevel?: 'LEVEL_1' | 'LEVEL_2' | 'LEVEL_3';

  /**
   * Apellido Materno de la Persona
   */
  apellidoMaterno?: string;

  /**
   * Apellido Paterno de la Persona
   */
  apellidoPaterno: string;

  /**
   * Celular de la Persona
   */
  celular?: string;

  /**
   * Tipo de Documento de Identidad
   */
  documentType?: 'DNI' | 'PASSPORT' | 'FOREIGN_ID' | 'HEALTH_CARD' | 'DRIVER_LICENSE' | 'SOCIAL_SECURITY' | 'OTHER';

  /**
   * Correo electrónico personal
   */
  email?: string;
  estadoCivil?: 'SOLTERO' | 'CASADO' | 'UNION_LIBRE' | 'SEPARADO' | 'DIVORCIADO' | 'VIUDO';

  /**
   * La fecha de nacimiento de la Persona
   */
  fechaNacimiento?: string;
  genero?: 'MASCULINO' | 'FEMENINO' | 'OTRO' | 'DESCONOCIDO';

  /**
   * Id de la Persona
   */
  id?: number;

  /**
   * Identificador Nacional de Salud
   */
  nationalHealthId?: string;

  /**
   * Identificador Nacional
   */
  nationalId?: string;

  /**
   * Nombre de la Persona
   */
  nombre: string;
  personTenantsIds?: Array<number>;

  /**
   * Tipo de Persona
   */
  personType?: 'DOCTOR' | 'NURSE' | 'PATIENT' | 'EMPLOYEE' | 'SUPER_ADMIN' | 'ADMIN' | 'RECEPTIONIST' | 'LAB_TECHNICIAN' | 'PHARMACIST';

  /**
   * Segundo nombre de la Persona
   */
  segundoNombre?: string;

  /**
   * Telefono de la Persona
   */
  telefono: string;

  /**
   * Datos del Tenant (Para vinculación o creación)
   */
  tenant?: Tenant;

  /**
   * Datos del Usuario asociado
   */
  user?: UserDto;
}